feat: refactor Google authentication logic
Inject GoogleAuthService into SocialAuthController 
to encapsulate user creation and social provider 
management. Update routes to apply API middleware 
to Google login endpoint for improved security.
